Fix issues with parsing of Service binary paths
Seatbelt parses and enumerates services via WMI, and attempts to pull information out of the binaries that the service points to. There were a few issues with the approach that was taken:

* It didn't cover cases where service paths were quoted or had spaces (such as Motorola's Bluetooh OBEX Service).
* It didn't handle cases where service binaries didn't have an extension (such as Carbon Black's Sensor agent).
* It didn't seem to correctly handle arguments passed to the service.
* When errors were encountered while parsing the file information, there was no try/catch around a single case, and hence the entire service parsing loop was exited (resulting in the rest of the services not being enumerated).

This PR attempts to fix this problem. It should handle quoted service paths, files without extensions, and cases where arguments are passed to the service exe.

Thoughts and feedback welcome. Thanks!